backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
Pillow>=10.0.0
psycopg2-binary>=2.9.0
